+++ source/ftrack_connect_cinema_4d/event.py
@@ -9,8 +9,12 @@ import os.path
 import c4d.gui
 
 import ftrack_connect_cinema_4d.asset
-import ftrack_connect_cinema_4d.core_message_event
 import ftrack_connect_cinema_4d.publish
+
+
+from ftrack_connect_cinema_4d.plugin.ftrack_message_data import (
+    FtrackMessageData
+)
 
 logger = logging.getLogger('ftrack_connect_cinema_4d.event')
 
@@ -29,7 +33,12 @@ def get_publish_options(session, data):
 
 def publish_media(session, options):
     logger.info(u'publish media')
-    result = ftrack_connect_cinema_4d.publish.publish(session, options)
+
+
+    result = FtrackMessageData.execute_in_main_thread(
+        ftrack_connect_cinema_4d.publish.publish, session, options
+    )
+
     logger.info(u'Published: {0}'.format(result))
     return result
 
@@ -37,18 +46,22 @@ def publish_media(session, options):
 def get_import_components(session, data):
     '''Export media with *options*.'''
     logger.info(u'Getting import components: {0!r}'.format(data))
-    return ftrack_connect_cinema_4d.asset.get_importable_components(
+    result = ftrack_connect_cinema_4d.asset.get_importable_components(
         session, data['versionId']
     )
+    logger.info(u'Result: {0!r}'.format(result))
+    return result
 
 
 def import_component(session, data):
     '''Import component in *data*.'''
     logger.info(u'Importing component: {0!r}'.format(data))
-    ftrack_connect_cinema_4d.core_message_event.send_event(
-        'IMPORT_OBJECT_FROM_FILE_PATH', data
+
+    result = FtrackMessageData.execute_in_main_thread(
+        ftrack_connect_cinema_4d.asset.import_object_from_file_path, *[], **data
     )
-    return True
+    logger.info(u'Result: {0!r}'.format(result))
+    return result
 
 
 #: Map functions to event names