+++ doc/release/release_notes.rst
@@ -7,7 +7,8 @@
 Release Notes
 *************
 
-.. release:: upcoming
+.. release:: 0.1.3
+    :date: 2017-06-28
 
     .. change:: fixed